VSCode compilation of ES32 and STM32 code

+++ Firmware/STM32/Common/configuration.h
@@ -18,7 +18,7 @@ public:
 	bool useStrands;
 	uint8_t strandFormat;
 
-	uint16_t elementCount;
+	uint8_t elementCount;
 	uint16_t elementWidth;
 	uint16_t elementHeight;
